requirements.txt
fastapi
uvicorn
python-multipart
python-docx
PyMuPDF
zhipuai
python-dotenv
